Rimosse entity vecchie, aggiunte entity + repository e impostazione case sentitive jspring

+++ src/backend/main/java/com/retrobased/market/repositories/RepositoryIndirizzoCliente.java
@@ -1,12 +1,10 @@
 package com.retrobased.market.repositories;
 
-import com.retrobased.market.entities.Cliente;
-import com.retrobased.market.entities.IndirizzoCliente;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RepositoryIndirizzoCliente extends JpaRepository<IndirizzoCliente,Integer> {
+public interface RepositoryIndirizzoCliente extends JpaRepository<IndirizzoCliente,Long> {
 
     boolean existsIndirizzoClienteByIdAndIdCliente(Long idIndirizzo, Cliente idCliente);
 }
